Refactor:Create single Logical operator for IndexApply
