+++ libgpopt/include/gpopt/operators/COperator.h
@@ -133,9 +133,8 @@ namespace gpopt
 				EopLogicalRename,
 				EopLogicalInnerApply,
 				EopLogicalInnerCorrelatedApply,
-				EopLogicalInnerIndexApply,
+				EopLogicalIndexApply,
 				EopLogicalLeftOuterApply,
-				EopLogicalLeftOuterIndexApply,
 				EopLogicalLeftOuterCorrelatedApply,
 				EopLogicalLeftSemiApply,
 				EopLogicalLeftSemiCorrelatedApply,
