+++ libgpopt/include/gpopt/operators/ops.h
@@ -85,9 +85,9 @@
 #include "gpopt/operators/CLogicalApply.h"
 #include "gpopt/operators/CLogicalInnerApply.h"
 #include "gpopt/operators/CLogicalInnerCorrelatedApply.h"
-#include "gpopt/operators/CLogicalInnerIndexApply.h"
+#include "gpopt/operators/CLogicalIndexApply.h"
 #include "gpopt/operators/CLogicalLeftOuterApply.h"
-#include "gpopt/operators/CLogicalLeftOuterIndexApply.h"
+//#include "gpopt/operators/CLogicalLeftOuterIndexApply.h"
 #include "gpopt/operators/CLogicalLeftOuterCorrelatedApply.h"
 #include "gpopt/operators/CLogicalLeftSemiApply.h"
 #include "gpopt/operators/CLogicalLeftSemiApplyIn.h"
